Rework: reworked commander commands to be in separate files instead of a big dump
