+++ bashlib/solid/src/shell/commands/shellcommands.ts
@@ -1,0 +1,15 @@
+import { addFetchCommand } from "./fetch";
+import { addListCommand } from './list';
+import { addCopyCommand } from './copy';
+import { addMoveCommand } from './mv';
+import { addRemoveCommand } from './remove';
+import { addFindCommand } from './find';
+import { addQueryCommand } from './query';
+import { addEditCommand } from './edit';
+import { addMkdirCommand } from './mkdir';
+import { addPermsCommand } from './perms';
+import { addShellCommand } from './shell';
+import { addTouchCommand } from './touch';
+import { addTreeCommand } from './tree';
+
+export { addFetchCommand, addListCommand, addCopyCommand, addMoveCommand, addRemoveCommand, addFindCommand, addQueryCommand, addEditCommand, addMkdirCommand, addPermsCommand, addShellCommand, addTouchCommand, addTreeCommand}
