+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "minimist": "^1.2.6"
+  }
+}
